chore: Move swarm notification types into a module
